+++ GameEngine/src/Actors/Player.cpp
@@ -1,9 +1,15 @@
 ﻿#include "Player.h"
 
+#include "../InputHandler/InputHandler.h"
 
-Player::Player(SDL_Rect* rect, const char filePath[], float collisionRadius, float speed, float2 direction) : Actor(rect, filePath, collisionRadius), IVelocity(speed, direction)
+
+Player::Player(
+    SDL_Rect* rect,
+    const char filePath[],
+    int collisionRadius,
+    float speed,
+    float2 direction) : Actor(rect, filePath, collisionRadius), IVelocity(speed, direction)
 {
-    
 }
 
 void Player::Fire()
@@ -18,5 +24,6 @@ void Player::Aim()
 
 void Player::Move(float2 input)
 {
+    
     
 }
