+++ GameEngine/src/Actors/Projectile.cpp
@@ -2,7 +2,7 @@
 
 #include "Projectile.h"
 
-Projectile::Projectile(SDL_Rect* rect, const char filePath[], float collisionRadius, float speed, float2 direction): Actor(rect, filePath, collisionRadius) ,IVelocity(speed, direction) {}
+Projectile::Projectile(SDL_Rect* rect, const char filePath[], int collisionRadius, float speed, float2 direction): Actor(rect, filePath, collisionRadius) ,IVelocity(speed, direction) {}
 
 void Projectile::Update(float deltaTime)
 {
