+++ GameEngine/src/Actors/Projectile.h
@@ -6,7 +6,7 @@
 class Projectile final : public Actor, public IVelocity
 {
 public:
-    Projectile(SDL_Rect* rect, const char filePath[], float collisionRadius, float speed, float2 direction);
+    Projectile(SDL_Rect* rect, const char filePath[], int collisionRadius, float speed, float2 direction);
     ~Projectile() override = default;
 
     void Update(float deltaTime) override;
